removed access to admin profile, stop password reset working, removed dark-mode stryling
i have stopped access to the admin profile page and return a 404 when someone trys to access this. I can add back later so its full functional for a demo if needed.
removed dark mode as it was annoying and i only want one set of styling for my page